+++ R/reexports.R
@@ -1,0 +1,45 @@
+##' @importFrom magrittr %>%
+##' @export
+magrittr::`%>%`
+
+##' @importFrom magrittr %<>%
+##' @export
+magrittr::`%<>%`
+
+
+##' @importFrom dplyr arrange
+##' @export
+dplyr::arrange
+
+##' @importFrom dplyr filter
+##' @export
+dplyr::filter
+
+##' @importFrom dplyr group_by
+##' @export
+dplyr::group_by
+
+##' @importFrom dplyr mutate
+##' @export
+dplyr::mutate
+
+##' @importFrom dplyr n
+##' @export
+dplyr::n
+
+##' @importFrom dplyr rename
+##' @export
+dplyr::rename
+
+##' @importFrom dplyr select
+##' @export
+dplyr::select
+
+##' @importFrom dplyr slice
+##' @export
+dplyr::slice
+
+##' @importFrom dplyr summarise
+##' @export
+dplyr::summarise
+